requirements.txt
langchain-community
faiss-cpu
sentence-transformers
langchain
pypdf 
streamlit
google.generativeai
